Add notification as parameter to onRemove
The spec of onRemove is the following

onRemove 	function 	null 	A callback function that will be called when the notification is about to be removed. The first argument is the original notification e.g. function (notification) { console.log(notification.title + 'was removed'); }

thus, the notification should be passed as parameter on line 31